add web apis and add city class

+++ SaveMyDate.Entities/Location.cs
@@ -3,8 +3,8 @@ namespace SaveMyDate.Entities
 {
     public class Location
     {
-        public decimal X { get; set; }
-        public decimal Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
         public Address Address { get; set; }
     }
 }
